Revert "feat: Personal view now shows any location and radiuses saved via fav…"
