+++ .env.development
@@ -2,6 +2,7 @@ EXTEND_ESLINT=true # Set linting to true on dev
 # developemnt API and keys
 REACT_APP_API_HOST='https://wmca-api-portal-staging.azure-api.net'
 REACT_APP_API_KEY='0d4cca4a2c5d40c3bfbbfe45d1bbf294'
-# Jon's test API
-# REACT_APP_API_HOST='https://transport-api-jm-dev.azure-api.net'
-# REACT_APP_API_KEY='b9fc95b60a2f4132af6756f81fbb2679'
+# Bus autocomplete API and keys
+REACT_APP_BUS_AUTOCOMPLETE_API='https://rtccdisruptionsbhrx2guer.azurewebsites.net'
+# Roads autocomplete key
+REACT_APP_ROADS_AUTOCOMPLETE_KEY='e0c1216f818a41be8d528ac1d4f7ebfd'
