+++ .env.production
@@ -1,3 +1,7 @@
 # Live API and keys
 REACT_APP_API_HOST='https://api.wmnetwork.co.uk'
 REACT_APP_API_KEY='514b8bafa009424289cb2d7a12c54f6e'
+# Bus autocomplete API and keys
+REACT_APP_BUS_AUTOCOMPLETE_API='https://rtccdisruptionsb5phceeso.azurewebsites.net'
+# Roads autocomplete key
+REACT_APP_ROADS_AUTOCOMPLETE_KEY='e0c1216f818a41be8d528ac1d4f7ebfd'
