+++ src/globalState/GlobalStateContext.tsx
@@ -3,9 +3,65 @@ import { useReducer } from 'preact/hooks';
 // Helpers
 import { getDisruptionCookieData, hasAnyFavourites, setCookie } from 'sharedHelpers';
 // Types
-import { RoadsFavEntity, TrainFavEntity } from 'sharedHelpers/cookies/types';
-import { ContextProviderProps, CreateContextProps } from 'sharedTypes';
-import { ActionType, InitialStateProps } from './types';
+import { TrainEntity } from 'sharedHelpers/cookies/types';
+import {
+  ContextProviderProps,
+  CreateContextProps,
+  DefaultModes,
+  DisruptionIndicatorTypes,
+} from 'sharedTypes';
+
+export type FavsStateProps = {
+  bus: DisruptionIndicatorTypes[];
+  tram: DisruptionIndicatorTypes[];
+  train: DisruptionIndicatorTypes[];
+  roads: DisruptionIndicatorTypes[];
+};
+
+export type InitialStateProps = {
+  editMode: boolean;
+  isRowExpandedOnMobile: {
+    bus: boolean;
+    tram: boolean;
+    train: boolean;
+    roads: boolean;
+  };
+  hasFavs: boolean;
+  favs: FavsStateProps;
+  previousFavs: FavsStateProps;
+  favsToRemoveOnSave: { mode: DefaultModes; id: string }[];
+};
+
+type ActionType =
+  | {
+      type: 'SET_EDIT_MODE';
+      payload: boolean;
+    }
+  | {
+      type: 'TOGGLE_ROW_VISIBILITY';
+      payload: { mode: DefaultModes; visible: boolean };
+    }
+  | {
+      type: 'UPDATE_SERVICES';
+      payload: {
+        mode: DefaultModes;
+        data: DisruptionIndicatorTypes[];
+      };
+    }
+  | {
+      type: 'REMOVE_SERVICE';
+      payload: {
+        mode: DefaultModes;
+        id: string;
+      };
+    }
+  | {
+      type: 'CANCEL_STATE';
+      payload: FavsStateProps;
+    }
+  | {
+      type: 'SAVE_NEW_STATE';
+    };
 
 const initialState: InitialStateProps = {
   editMode: false,
@@ -100,23 +156,13 @@ export const GlobalContextProvider = ({ children }: ContextProviderProps): JSX.E
           if (!cookieObj || !cookieObj.favs) return; // No fav object in cookies, then escape out of function
           const cookiesFavsObj = cookieObj.favs;
 
-          const cookiesModeArr = cookiesFavsObj[mode] as TrainFavEntity[] &
-            RoadsFavEntity[] &
-            string[]; // set as both because TypeScript has a bug with filtering union arrays
+          const cookiesModeArr = cookiesFavsObj[mode] as TrainEntity[] & string[]; // set as both because TypeScript has a bug with filtering union arrays
 
           if (!cookiesModeArr) return; // Avoid any undefined or empty arrays
 
           // Get round the TypeScript filter union array bug by splitting our filters into two seperate calls via an if statement and explicitely stating the type of item we expect to be here
           if (mode === 'train') {
-            cookieObj.favs[mode] = cookiesModeArr.filter(
-              (item: TrainFavEntity) => item.line !== id
-            ); // Map filtered array back to our cookieObj
-          } else if (mode === 'roads') {
-            // As we set our id to be address + radius in disruption indicator for roads, we need to check if the address + radius is the same as our id. If it is, it will return false (meaning we filter it out). Otherwise return true and keep that cookie.
-            cookieObj.favs[mode] = cookiesModeArr.filter(({ address, radius }: RoadsFavEntity) => {
-              if (address && radius) return address + radius !== id;
-              return true;
-            });
+            cookieObj.favs[mode] = cookiesModeArr.filter((item: TrainEntity) => item.line !== id); // Map filtered array back to our cookieObj
           } else {
             cookieObj.favs[mode] = cookiesModeArr.filter((item: string) => item !== id); // Map filtered array back to our cookieObj
           }
